brain/pyproject.toml
[project]
name = "bmo-brain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Kevin Gutierrez Gomez", email = "kevingtzz@outlook.com" }
]
requires-python = ">=3.11"
dependencies = [
    "langchain-core",
    "langchain-openai>=1.1.10",
    "langgraph",
    "python-dotenv>=1.2.1",
    "websockets",
]

[project.scripts]
bmo-brain = "bmo_brain:main"

[build-system]
requires = ["uv_build>=0.10.2,<0.11.0"]
build-backend = "uv_build"
